move project away from typescript
