+++ packages/frontendmentor/fem-password-generator-app/src/Checkbox.jsx
