+++ packages/frontendmentor/fem-password-generator-app/src/InputCard.jsx
