+++ packages/frontendmentor/fem-password-generator-app/src/OutputCard.jsx
