+++ packages/frontendmentor/fem-password-generator-app/src/Slider.jsx
