+++ packages/frontendmentor/fem-password-generator-app/src/utils/PasswordGenerator.jsx
